+++ db/sqlc/report.sql.go
@@ -21,7 +21,7 @@ INSERT INTO report_forms (
 
 type CreateReportParams struct {
 	Comment           string `json:"comment"`
-	ProfessorRatingID int64  `json:"professorRatingID"`
+	ProfessorRatingID int32  `json:"professorRatingID"`
 	UserID            string `json:"userID"`
 }
 
@@ -93,7 +93,7 @@ RETURNING id, comment, status, request_date, verified_date, professor_rating_id,
 
 type UpdateReportParams struct {
 	Status Statusrequest `json:"status"`
-	ID     int64         `json:"id"`
+	ID     int32         `json:"id"`
 }
 
 func (q *Queries) UpdateReport(ctx context.Context, arg UpdateReportParams) (ReportForm, error) {
